requirements.txt
Django>=1.6,<1.7
South==0.8.4
Whoosh==2.6.0
django-bootstrap3==4.11.0
django-haystack==2.2.0
django-mptt==0.6.1
django-taggit==0.12
httplib2==0.9
lxml==3.3.5
pep8==1.5.7
requests==2.3.0
